@@ -4,16 +4,16 @@
     } else if (typeof exports === 'object') {
         module.exports = factory(require('underscore'), require('simple-statistics'));
     } else {
-        root.computeBoxplotStats = factory(root._, root.simpleStatistics);
+        root.computeBoxplotStats = factory(root._, root.ss);
     }
-}(this, function (_, simpleStatistics) {
+}(this, function (_, ss) {
 
     var computeBoxplotStats = function (data) {
 
-        var median = simpleStatistics.median(data);
+        var median = ss.median(data);
 
-        var lowerQuartile = simpleStatistics.quantile(data, 0.25),
-            upperQuartile = simpleStatistics.quantile(data, 0.75),
+        var lowerQuartile = ss.quantile(data, 0.25),
+            upperQuartile = ss.quantile(data, 0.75),
             interQuartileRange = upperQuartile - lowerQuartile;
 
         var lowerOutlierCutoff = lowerQuartile - 1.5 * interQuartileRange,
@@ -28,15 +28,15 @@
             }
         });
 
-        var minimum = simpleStatistics.min(nonOutliers),
-            maximum = simpleStatistics.max(nonOutliers);
+        var minimum = ss.min(nonOutliers),
+            maximum = ss.max(nonOutliers);
 
         return {
-            minimum: minimum,
-            lowerQuartile: lowerQuartile,
-            median: median,
-            upperQuartile: upperQuartile,
-            maximum: maximum,
+            whiskerLow: minimum,
+            quartile1: lowerQuartile,
+            quartile2: median,
+            quartile3: upperQuartile,
+            whiskerHigh: maximum,
             outliers: outliers,
         };
     };
